requirements.txt
aiogram==3.2.0
dispatcher==1.0
python-dotenv
psycopg2-binary